Simplification and standardization of how word counts are handled.
AbstractProjectViewer now has a "chapterCounts" set that holds an instance of ChapterCounts against a Chapter.

The counts are inited when the project is opened to give the starting word counts, then refreshed
whenever a change is made to a chapter (see: AbstractEditorPanel).

Calculation and storage of the A4 page count has been moved out of ChapterCounts and is handled separately, see:
AbstractProjectViewer.getChapterA4PageCount(Chapter) and AbstractProjectViewer.getAllChaptersA4PageCount().