A new way of doing updates has been introduced, this is to support the new way of performing the updates.  From version 2.5.4 onwards an executable file will be run once QW has exited to perform the update.  This is instead of using a zip file and updating particular jar files.
A new interface QuollWriterUpdater has been added to allow for custom install methods.  DefaultQuollWriterUpdater has been added
to handle the new update method.  GetLatestVersion is now deprecated and no longer used, it will be removed in a future release.

A new property "quollWriterUpdaterClassName" has been added to allow configuration of the class to use for updating.  Implementations
must have a zero arg public constructor since new instances are created via Class.newInstance.

If you need to provide a custom installation method then implement QuollWriterUpdater and set the class name in the property.

You can now add Runnables to the environment that are run just before QW is shutdown, the DefaultQuollWriterUpdater makes use of
this to execute the new install file on exit.