Changes:
The use of java.util.Timer has been changed to now use a java.util.concurrent.ScheduledThreadPoolExecutor with a custom (but vanilla) thread factory.
Timer kept being the source of some strange illegal state exceptions that were next to impossible to track down.  This change also removes the use of TimerTasks as the units of work, replaced by Runnables.  The old methods are still present but deprecated and will be removed in a future release (probably the next release).

All threads are created with lowest priority and are NOT run on the event thread so ensure that any UI updating is performed via a call to UIUtils.doLater.  The runnables also are not try/catch wrapped so you MUST do this yourself.  This decision is to allow you to unschedule the runnables if needed.

The general look of the runnable should be:

new Runnable ()
{

    @Override
    public void run ()
    {

        try
        {

            // Do some background work.
            UIUtils.doLater (new ActionListener ()
            {

                @Override
                public void actionPerformed (ActionEvent ev)
                {

                    // Update the UI.

                }

            });

        } catch (Exception e) {

            Environment.logError ("Error", e);

        }

    }

}