A partial commit that starts off the work of allowing the text used in the UI to be customized, specifically for other languages.
A new class "com.quollwriter.LanguageStrings" has been added that manages access to the relevant string with constant identifiers providing indexing into the text.

The default strings themselves are held in "data/default-ui-language-strings.json".  This file will change heavily soon.  It is in a JSON format.

The defacto structuring provides heirarchical access, this was deemed preferable to the simplistic and repetitive format suggested by Java itself (i.e. a preference file).  The JSON file can easily be flattened into a .csv for easy user modification/creation.

Access to the relevant string in the Java code is provided through a number of new "getUIString" methods in "com.quollwriter.Environment".  The access is through a number of constants from LanguageStrings.

Thus a structure in the JSON file such as:

{
    "mytext" :
    {
        "buttons" :
        {
            "help" :
            {
                "tooltip" : "Hello World!"
            }
        }
    }

}

Would be accessed with a call such as:

Environment.getUIString (LanguageStrings.mytext, LanguageStrings.buttons, LanguageStrings.help, LanguageStrings.tooltip);

This mechanism allows for heavy reuse of string constants.

Values within the JSON file can reference other values through the use of a ${a.b.c...} notation, thus to reference the tooltip above in another structure the following would be used:

{
    "myreference" : "${mytext.buttons.help.tooltip}"
}

Standard object name values are referenced using the usual notation "{a}", for example "{chapter}".

Sometimes values within the text are replaced with dynamic values in the code.  These are referenced using the standard notation expected by java.lang.String.format, i.e. %1s, %2s, ...

Putting all this together you might have a string:

{
    "mytitle" : "${mytext.buttons.help.tooltip} %1s {chapter} %2s"
}

The following special values are used to comment the values and aid with validation of the text.

* :comment[.name] - Used to provide a comment about a particular field or section.  These values will be displayed to the person providing the translation or other strings file.  If a string contains %xs values then the comment will describe what each replacement is for.  For example:

{
    "mysection" :
    {
        ":comment" : "This is my section.  It does stuff.",
        "text" : "This is my %1s.",
        ":comment.text" : "%1s is the name of something here."
    }
}

* :scount.name - Used to indicate how many %xs values are expected in the string.